+++ fabricpc/training/train_autoregressive.py
@@ -22,7 +22,7 @@ import optax
 
 from fabricpc.core.types import GraphParams, GraphState, GraphStructure
 from fabricpc.core.inference import run_inference
-from fabricpc.graph_initialization.graph_net import compute_local_weight_gradients
+from fabricpc.core.learning import compute_local_weight_gradients
 from fabricpc.graph_initialization.state_initializer import initialize_graph_state
