+++ fabricpc/utils/helpers.py
@@ -1,4 +1,5 @@
-import jax.numpy as jnp
+from typing import Dict
+from jax import numpy as jnp
 from fabricpc.core.types import GraphState
 
 
@@ -34,3 +35,26 @@ def layernorm(
     mean = jnp.mean(x, axis=-1, keepdims=True)
     variance = jnp.var(x, axis=-1, keepdims=True)
     return gamma * (x - mean) / jnp.sqrt(variance + eps) + beta
+
+
+# TODO call this method to reduce boilerplate code for clamping latents in training loops.
+def set_latents_to_clamps(
+    state: GraphState,
+    clamps: Dict[str, jnp.ndarray],
+) -> GraphState:
+    """
+    Set the latent states of specified nodes to their clamped values.
+
+    Args:
+        state: Current graph state
+        clamps: Dictionary of clamped values, keyed on node names
+
+    Returns:
+        Updated GraphState with latents set to clamped values
+    """
+    for node_name, clamp_value in clamps.items():
+        if node_name in state.nodes:
+            state = update_node_in_state(
+                state, node_name, z_latent=jnp.asarray(clamp_value, dtype=jnp.float32)
+            )
+    return state
